+++ index.js
@@ -1,3 +1,0 @@
-const ClusterConfiguration = require('./src/ClusterConfiguration');
-
-module.exports = new ClusterConfiguration();