+++ index.ts
@@ -1,0 +1,3 @@
+import { ClusterConfiguration } from './src/ClusterConfiguration';
+
+export default new ClusterConfiguration();