+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # [0.1.0](https://github.com/kapetacom/local-cluster-config/compare/v0.0.23...v0.1.0) (2023-05-08)
 
-
 ### Features
 
-* add getDockerConfiguration to ClusterConfig [KAP-609] ([f7587db](https://github.com/kapetacom/local-cluster-config/commit/f7587db0b15557a56f94bff021b75bf9539724a5))
+-   add getDockerConfiguration to ClusterConfig [KAP-609] ([f7587db](https://github.com/kapetacom/local-cluster-config/commit/f7587db0b15557a56f94bff021b75bf9539724a5))